update to the trivia api in __init__ and test_flaskr

+++ starter/backend/flaskr/__init__.py
@@ -1,8 +1,11 @@
+from logging import FATAL
 import os
 from flask import Flask, json, request, abort, jsonify
 from flask_sqlalchemy import SQLAlchemy
 from flask_cors import CORS
 import random
+from sqlalchemy.sql.base import NO_ARG
+from sqlalchemy.sql.elements import Null
 
 from sqlalchemy.sql.expression import select
 
@@ -10,6 +13,7 @@ from models import setup_db, Question, Category
 
 QUESTIONS_PER_PAGE = 10
 
+# Displays only 10 questions per page
 def paginate_question(request, selection):
   page = request.args.get('page', 1, type=int)
   start = (page - 1) * QUESTIONS_PER_PAGE
@@ -21,32 +25,24 @@ def paginate_question(request, selection):
   return current_questions
 
 def create_app(test_config=None):
-  # create and configure the app
+  # Creates and configure the app
   app = Flask(__name__)
   setup_db(app)
-  
-  '''
-  @TODO: Set up CORS. Allow '*' for origins. Delete the sample route after completing the TODOs
-  '''
+  # Setups CORS to allow Cross Origin Resource Sharing
   CORS(app)
 
-  '''
-  @TODO: Use the after_request decorator to set Access-Control-Allow
-  '''
   @app.after_request
   def after_request(response):
     response.headers.add('Access-Control-Allow-Headers', 'Content-Type,Authorization,true')
     response.headers.add('Access-Control-Allow-Methods', 'GET,PUT,POST,DELETE,OPTIONS')
     
     return response
-  '''
-  @TODO: 
-  Create an endpoint to handle GET requests 
-  for all available categories.
-  '''
+  
+  # Gets the categories from the database
   @app.route('/categories')
   def retrieve_categories():
     categories = Category.query.order_by(Category.id).all()
+    # Formats the categories into a dictionary {id:type}
     formatted_categories = {category.id: category.type for category in categories}
 
     return jsonify({
@@ -55,63 +51,52 @@ def create_app(test_config=None):
       'total_categories': len(categories)
     })
 
-  '''
-  @TODO: 
-  Create an endpoint to handle GET requests for questions, 
-  including pagination (every 10 questions). 
-  This endpoint should return a list of questions, 
-  number of total questions, current category, categories. 
-
-  TEST: At this point, when you start the application
-  you should see questions and categories generated,
-  ten questions per page and pagination at the bottom of the screen for three pages.
-  Clicking on the page numbers should update the questions. 
-  '''
+  # Gets all questions from the database 
   @app.route('/questions')
   def retrieve_questions():
+    # Gets all questions
     selection = Question.query.order_by(Question.id).all()
+    # Gets only 10 questions per page
     current_questions = paginate_question(request, selection) 
+    # Gets all categories
     categories = Category.query.order_by(Category.id).all()
+    # Formats the categories into a dictionary {id:type}
     formatted_categories = {category.id: category.type for category in categories}
 
+    if len(current_questions) == 0:
+      abort(404)
+
     return jsonify({
+      'success': True,
       'questions': current_questions,
       'total_questions': len(selection),
       'categories': formatted_categories,
       'current_category': None
     })
 
-  '''
-  @TODO: 
-  Create an endpoint to DELETE question using a question ID. 
-
-  TEST: When you click the trash icon next to a question, the question will be removed.
-  This removal will persist in the database and when you refresh the page. 
-  '''
+  # Delete a question by its id and after deleting displays the questions updated
   @app.route('/questions/<int:question_id>/delete', methods=['DELETE'])
   def delete_question(question_id):
-    question = Question.query.filter(Question.id==question_id).one_or_none()
-    question.delete()
+    try:
+      question = Question.query.filter(Question.id==question_id).one_or_none()
+      
+      if question is None:
+        abort(404)
+      
+      question.delete()
 
-    questions = Question.query.order_by(Question.id).all()
+      questions = Question.query.order_by(Question.id).all()
 
-    return jsonify({
-      'success': True,
-      'deleted': question_id,
-      'questions': paginate_question(request, questions),
-      'total_questions': len(questions)
-    })
-
-  '''
-  @TODO: 
-  Create an endpoint to POST a new question, 
-  which will require the question and answer text, 
-  category, and difficulty score.
-
-  TEST: When you submit a question on the "Add" tab, 
-  the form will clear and the question will appear at the end of the last page
-  of the questions list in the "List" tab.  
-  '''
+      return jsonify({
+        'success': True,
+        'deleted': question_id,
+        'questions': paginate_question(request, questions),
+        'total_questions': len(questions)
+      })
+    except:
+      abort(422)
+  
+  # Creates a new question and add into the database 
   @app.route('/questions', methods=['POST'])
   def create_question():
     question = request.get_json()['question']
@@ -119,99 +104,109 @@ def create_app(test_config=None):
     category = request.get_json()['category']
     difficulty = request.get_json()['difficulty']
     
-    question_item = Question(question=question, answer=answer, category=category, difficulty=difficulty)
-    question_item.insert()
+    try:
+      question_item = Question(question=question, answer=answer, category=category, difficulty=difficulty)
+      question_item.insert()
 
-    return jsonify({
-      'success': True,
-      'created': question_item.id,
-    })
-    
-  '''
-  @TODO: 
-  Create a POST endpoint to get questions based on a search term. 
-  It should return any questions for whom the search term 
-  is a substring of the question. 
-
-  TEST: Search by any phrase. The questions list will update to include 
-  only question that include that string within their question. 
-  Try using the word "title" to start. 
-  '''
+      return jsonify({
+        'success': True,
+        'created': question_item.id,
+      })
+    except:
+      abort(422)
+  
+  # Searches for a question by its searchTerm passed by json from the search form 
   @app.route('/questions/search', methods=['POST'])
   def search_questions():
-    search = request.get_json()['searchTerm']
-    selection = Question.query.order_by(Question.id).filter(Question.question.ilike('%{}%'.format(search)))
-    current_questions = paginate_question(request, selection)
+    try:
+      # Gets the searchTerm from the form
+      search = request.get_json()['searchTerm']
+      # Gets all questions that have the same searchTerm in the title
+      selection = Question.query.order_by(Question.id).filter(Question.question.ilike('%{}%'.format(search)))
+      # Format the questions to show only 10 per page
+      current_questions = paginate_question(request, selection)
 
-    return jsonify({
-      'success': True,
-      'questions': current_questions,
-      'total_questions': len(selection.all())
-    })
-
-  '''
-  @TODO: 
-  Create a GET endpoint to get questions based on category. 
-
-  TEST: In the "List" tab / main screen, clicking on one of the 
-  categories in the left column will cause only questions of that 
-  category to be shown. 
-  '''
+      return jsonify({
+        'success': True,
+        'questions': current_questions,
+        'total_questions': len(selection.all()),
+        'current_category': None
+      })
+    except:
+      abort(422)
+  
+  # Gets the questions by its category, only displays questions with the same category
   @app.route('/categories/<int:category_id>/questions')
   def questions_by_category(category_id):
-    selection = Question.query.order_by(Question.id).filter(Question.category == category_id).all()
-    current_questions = paginate_question(request, selection)
+    try:
+      selection = Question.query.order_by(Question.id).filter(Question.category == category_id).all()
+      current_questions = paginate_question(request, selection)
+      current_category = Category.query.filter(Category.id == category_id).one_or_none()
+      category = current_category.format()
 
-    return jsonify({
-      'success': True,
-      'questions': current_questions
-    })
-
-  '''
-  @TODO: 
-  Create a POST endpoint to get questions to play the quiz. 
-  This endpoint should take category and previous question parameters 
-  and return a random questions within the given category, 
-  if provided, and that is not one of the previous questions. 
-
-  TEST: In the "Play" tab, after a user selects "All" or a category,
-  one question at a time is displayed, the user is allowed to answer
-  and shown whether they were correct or not. 
-  '''
+      return jsonify({
+        'success': True,
+        'questions': current_questions,
+        'total_questions': len(selection),
+        'current_category': [value for key, value in category.items()][1]
+      })
+    except:
+      abort(422)
+  
+  # Gets the questions from the database randomly to play the quiz
   @app.route('/quizzes', methods=['POST'])
   def play():
+    # Gets the quiz category and the previous questions to not repeat them
     quiz_category = request.get_json()['quiz_category']
     previous_questions = request.get_json()['previous_questions']
 
-    category_id = int(quiz_category['id']) + 1
+    try:
+      category_id = int(quiz_category['id']) + 1
 
-    if quiz_category['id'] == 0:
-      quiz = Question.query.all()
-    
-    else:
-      quiz = Question.query.filter(Question.category == category_id).all()
+      # Gets the questions by the category selected
+      if quiz_category['id'] == 0:
+        quiz = Question.query.all()
+      
+      else:
+        quiz = Question.query.filter(Question.category == category_id).all()
 
-    question_selected = []
+      question_selected = []
 
-    for question in quiz:
-      if question.id not in previous_questions: 
-        question_selected.append(question.format())
+      # Checks if the question have not been picked yet
+      for question in quiz:
+        if question.id not in previous_questions: 
+          question_selected.append(question.format())
 
-    if len(question_selected) != 0:
-      result = random.choice(question_selected)
-      return jsonify({
-        'question': result
-      })
-    else:
-      return jsonify({
-        'question': False
-      })
+      # Selects a random question from the list of questions 
+      if len(question_selected) != 0:
+        result = random.choice(question_selected)
+        return jsonify({
+          'question': result
+        })
+      else:
+        return jsonify({
+          'question': False
+        })
+    except:
+      abort(422)
 
-  '''
-  @TODO: 
-  Create error handlers for all expected errors 
-  including 404 and 422. 
-  '''
+  # Error handlers 
+  @app.errorhandler(404)
+  def not_found(error):
+    return jsonify({
+      'success': False,
+      'error': 404,
+      'message': 'Resource not found'
+    }), 404
+  
+  
+  @app.errorhandler(422)
+  def unprocessable(error):
+    return jsonify({
+      'success': False,
+      'error': 422,
+      'message': 'Unprocessable'
+    }), 422
   
   return app
